Refactor (shrunk) m1 ablation code
