+++ QuantaTextToSql/ablate/__init__.py
@@ -1,9 +1,6 @@
-from .ablate_bm1 import (
+from .ablate_m1 import (
     load_bm1,
-    collect_bm1_activations,
-    ablate_bm1)
-
-from .ablate_tm1 import (
     load_tm1,
-    collect_tm1_activations,
-    ablate_tm1)
+    collect_m1_activations,
+    ablated_m1_inference,
+)
