+++ QuantaTextToSql/ablate/ablate_m1.py
@@ -5,6 +5,18 @@ from QuantaTextToSql.training_data.generate_utils import output_inference_text
 
 
 # Load the tokenizer and base model from Hugging Face
+def load_bm1():
+    tokenizer = AutoTokenizer.from_pretrained("roneneldan/TinyStories-Instruct-2Layers-33M")
+    model = AutoModelForCausalLM.from_pretrained("roneneldan/TinyStories-Instruct-2Layers-33M")
+    
+    # Ensure the tokenizer has a pad_token
+    if tokenizer.pad_token is None:
+        tokenizer.pad_token = tokenizer.eos_token 
+        
+    return tokenizer, model
+
+
+# Load the tokenizer and base model from Hugging Face
 def load_tm1():
     auth_token = os.getenv("HF_AUTH_TOKEN")
 
@@ -24,16 +36,15 @@ def load_tm1():
     return tokenizer, model
 
 
-# Dictionary to store average activations for all layers, heads, and MLPs
-average_tm1_activations = {
-    "head": {},  # Dict of layer -> list of average activations per head
-    "mlp": [],   # List of MLP average activations
-    "layer": []  # List of layer average activations
-}
-
-
 # Function to collect average activations for all heads, MLPs, and layers
-def collect_tm1_activations(model, tokenizer, batched_inputs):
+def collect_m1_activations(model, batched_inputs):
+
+    # Dictionary to store average activations for all layers, heads, and MLPs
+    cached_avg_activations = {
+        "head": {},  # Dict of layer -> list of average activations per head
+        "mlp": [],   # List of MLP average activations
+        "layer": []  # List of layer average activations
+    }
 
     # Hook to collect average activations
     def collect_activations_hook(module, input, output, layer_index=None, head_index=None):
@@ -43,15 +54,15 @@ def collect_tm1_activations(model, tokenizer, batched_inputs):
         if layer_index is not None:
             if head_index is not None:
                 # Store average activation for a specific attention head from output[BatchSize=100, SeqLen=183, HiddenDimension=1024]
-                if layer_index not in average_tm1_activations["head"]:
-                    average_tm1_activations["head"][layer_index] = []
-                while len(average_tm1_activations["head"][layer_index]) <= head_index:
-                    average_tm1_activations["head"][layer_index].append(None)
+                if layer_index not in cached_avg_activations["head"]:
+                    cached_avg_activations["head"][layer_index] = []
+                while len(cached_avg_activations["head"][layer_index]) <= head_index:
+                    cached_avg_activations["head"][layer_index].append(None)
                 head_activation = output[:, :, head_index].mean(dim=0).detach().clone() # Gives head_activation[SeqLen=183]
-                average_tm1_activations["head"][layer_index][head_index] = head_activation
+                cached_avg_activations["head"][layer_index][head_index] = head_activation
             else:
                 # Store average activation for the entire layer
-                average_tm1_activations["layer"].append(output.mean(dim=0).detach().clone())
+                cached_avg_activations["layer"].append(output.mean(dim=0).detach().clone())
 
     num_heads = model.config.num_attention_heads
     
@@ -69,8 +80,8 @@ def collect_tm1_activations(model, tokenizer, batched_inputs):
         # Collect MLP activations by averaging the MLP output directly
         def collect_mlp_output(module, input, output, li=layer_index):
             mlp_activation = output.mean(dim=0).detach().clone() # gives size [SeqLen=183, HiddenDimension=1024]
-            #print(f"collect_tm1: output.shape: {output.shape}, mlp_activation.shape: {mlp_activation.shape}") 
-            average_tm1_activations["mlp"].append(mlp_activation)
+            #print(f"collect_bm1: output.shape: {output.shape}, mlp_activation.shape: {mlp_activation.shape}") 
+            cached_avg_activations["mlp"].append(mlp_activation)
 
         
         layer.mlp.register_forward_hook(collect_mlp_output)
@@ -78,9 +89,11 @@ def collect_tm1_activations(model, tokenizer, batched_inputs):
     # Run inference (forward pass) without ablation to collect all average activations
     model(**batched_inputs)
 
+    return cached_avg_activations
+
 
 # Function to ablate using pre-collected average activations
-def ablate_tm1(tokenizer, model, batched_inputs, node_type="layer", layer_index=0, head_index=None):
+def ablated_m1_inference(tokenizer, model, cached_avg_activations, batched_inputs, node_type="layer", layer_index=0, head_index=None):
     
     # Ablation hook that uses pre-stored average values
     def ablation_hook(module, input, output):
@@ -88,16 +101,15 @@ def ablate_tm1(tokenizer, model, batched_inputs, node_type="layer", layer_index=
             output = output[0]
         
         # Replace output with average activation depending on node type
-        # Replace output with average activation depending on node type
         if node_type == "head" and head_index is not None:
-            the_activation = average_tm1_activations["head"][layer_index][head_index].clone()
+            the_activation = cached_avg_activations["head"][layer_index][head_index].clone()
             the_activation = the_activation.repeat(output.size(0), 1)
             
             #print(f"ablate_head: output.shape: {output.shape}, the_activation.shape: {the_activation.shape}")
             output[:, :, head_index] = the_activation
 
         elif node_type == "mlp":
-            the_activation = average_tm1_activations["mlp"][layer_index].clone()
+            the_activation = cached_avg_activations["mlp"][layer_index].clone()
             the_activation = the_activation.unsqueeze(0)
 
             # Adjust both sequence length (dim 1) and batch size (dim 0)
@@ -113,9 +125,9 @@ def ablate_tm1(tokenizer, model, batched_inputs, node_type="layer", layer_index=
 
             assert the_activation.shape == output.shape, f"Shape mismatch: activation {the_activation.shape} vs output {output.shape}"
             output[:] = the_activation
-            
+
         elif node_type == "layer":
-            the_activation = average_tm1_activations["layer"][layer_index].clone()
+            the_activation = cached_avg_activations["layer"][layer_index].clone()
             the_activation = the_activation.unsqueeze(0)
 
             # Adjust both sequence length (dim 1) and batch size (dim 0)
