+++ tests/test_suite.py
@@ -4,8 +4,7 @@ from tests.test_fragments import TestFragments
 from tests.test_cs1 import TestCommandSet1
 from tests.test_cs2 import TestCommandSet2
 from tests.test_cs3 import TestCommandSet3
-from tests.test_ablate_bm1 import TestAblateBM1
-from tests.test_ablate_tm1 import TestAblateTM1
+from tests.test_ablate_m1 import TestAblateBM1, TestAblateTM1
 
 
 if __name__ == '__main__':
